% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KostraApply.R
\name{KostraApply}
\alias{KostraApply}
\title{Apply a function with data frame output over data subsets}
\usage{
KostraApply(data, by, Fun, ...)
}
\arguments{
\item{data}{Data frame}

\item{by}{Vector of variable names (or variable numbers) defining the subsets or alternatively a named list or data frame (see examples).}

\item{Fun}{Function to be applied on data subsets}

\item{...}{Arguments passed to Fun}
}
\value{
A data.frame
}
\description{
Funksjonen er kopiert fra Kostra-pakken
}
\note{
The output ordering of the example depends on default.stringsAsFactors()
}
\examples{
1+1

}
