% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_match.R
\name{lower_match}
\alias{lower_match}
\title{Navn i data som er matchende etter \code{\link{tolower}}}
\usage{
lower_match(x, code)
}
\arguments{
\item{x}{Data eller bare vektor med navn}

\item{code}{Kode det skal matches mot}
}
\value{
Som input-\code{code}, eventuelt endret  p.g.a. første match som ble funnet.
}
\description{
Navn i data som er matchende etter \code{\link{tolower}}
}
\examples{
z <- SdcData("syssel27")
lower_match(z, "frtk_id_ssb")
lower_match(z, "frtk_ID_ssb")
lower_match(z, "ALDER")  # ingen match funnet 
lower_match(z, "ALDER6")
lower_match(z, "Sektor")
}
