% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{Make_FRTK_VIRK_UNIK_AggVar}
\alias{Make_FRTK_VIRK_UNIK_AggVar}
\title{Genererer variabel som er foretak-ID dersom den finnes, ellers virksomhet-ID, eller en unik ID.}
\usage{
Make_FRTK_VIRK_UNIK_AggVar(
  data,
  frtk = "FRTK_ID_SSB",
  virk = "VIRK_ID_SSB",
  unik = "UNIK_ID",
  makeunik = TRUE,
  varnames = c("FRTK_VIRK_UNIK", "sysselsatte", "sysselsatteVIRK", "sysselsatteFRTK",
    "antVIRK", "antFRTK", "ant_FRTK_VIRK_UNIK"),
  miss = "",
  dupvar = NULL
)
}
\arguments{
\item{data}{data frame}

\item{frtk}{foretak variabel}

\item{virk}{virksomhet variabel}

\item{unik}{unik variabel}

\item{makeunik}{Unik variabel genereres ved TRUE eller antas det at den finnes}

\item{varnames}{Vektor av 7 variabelnavn. Bruk NA dersom variabel ikke ønskes i output.}

\item{miss}{Kode for missing}

\item{dupvar}{Variabler som definerer grupper}
}
\value{
input data med nye variabler
}
\description{
Genererer variabel som er foretak-ID dersom den finnes, ellers virksomhet-ID, eller en unik ID.
}
\examples{
z <- OylData("syssel7")
Make_FRTK_VIRK_UNIK_AggVar(z)
Make_FRTK_VIRK_UNIK_AggVar(z, 
    varnames = c("FRTK_VIRK_UNIK", "sysselsatte", NA, NA, NA, NA, "ant_FRTK_VIRK_UNIK"))
z$id <- 1:7
Make_FRTK_VIRK_UNIK_AggVar(z, unik = "id", makeunik = FALSE)
}
