% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{MissingReplaceFirst}
\alias{MissingReplaceFirst}
\title{First column with missing values replaced}
\usage{
MissingReplaceFirst(x, missing = NULL)
}
\arguments{
\item{x}{data.frame}

\item{missing}{Values treated as missing}
}
\value{
First column with missing values replaced
}
\description{
The first column is returned with missing values sequentially replaced with values from the other columns
}
\examples{
a <- data.frame(x = c("yes", "", "no", ""), y = c("", "ok", "ok", ""), z = "z")
print(a)
MissingReplaceFirst(a)
MissingReplaceFirst(a, "")
MissingReplaceFirst(a, c("", "yes"))
a[3, 1] <- NA
MissingReplaceFirst(a)
}
