% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{Ones2agg}
\alias{Ones2agg}
\title{Ones to be aggregated to counts}
\usage{
Ones2agg(data, var, miss = NULL, dupvar = NULL, int = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{var}{A variable in data}

\item{miss}{Values treated as missing}

\item{dupvar}{variables defining groups (see details)}

\item{int}{Integers returned when TRUE}
}
\value{
Vector of ones and zeros
}
\description{
Based on a variable in the data, missing and duplicates within groups are set to zero.
}
\details{
To treat duplicates without grouping set \code{dupvar} to the same value as \code{var}.
Also \code{character(0)} or \code{integer(0)} will work when \code{var} is specified
by name or number, respectively.
}
\examples{
a <- data.frame(x = rep(c("A", "B", "C", ""), each = 2), 
                y = c(5, 5, 5, 6, 6, 6, 7, 7), 
                z = c("r", rep("s", 7)))
cbind(a, ones = Ones2agg(a, "x"))
cbind(a, ones = Ones2agg(a, "x", ""))
cbind(a, ones = Ones2agg(a, "x", "", "y"))
cbind(a, ones = Ones2agg(a, "x", "", "z"))
cbind(a, ones = Ones2agg(a, "x", "", c("y", "z")))
cbind(a, ones = Ones2agg(a, "x", dupvar = "x"))
cbind(a, ones = Ones2agg(a, "x", dupvar = character(0)))
a[1, 1] <- NA
cbind(a, ones = Ones2agg(a, "x"))
}
