% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OylData.R
\name{OylData}
\alias{OylData}
\title{Function that returns a dataset}
\usage{
OylData(dataset)
}
\arguments{
\item{dataset}{Name of data set within the SdcForetakPerson package}
}
\value{
data frame
}
\description{
Function that returns a dataset
}
\examples{
OylData("syssel7")
z <- OylData("syssel100")
}
\author{
Øyvind Langsrud
}
