% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SdcData.R
\name{SdcData}
\alias{SdcData}
\title{Function that returns a dataset}
\usage{
SdcData(dataset)
}
\arguments{
\item{dataset}{Name of data set within the SdcForetakPerson package}
}
\value{
data frame
}
\description{
Function that returns a dataset
}
\examples{
SdcData("syssel7")
z <- SdcData("syssel100")
}
\author{
Øyvind Langsrud
}
