% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SdcForetakPerson.R
\name{SdcForetakPerson}
\alias{SdcForetakPerson}
\title{Prikking av foretak og avrunding eller prikking av personer}
\usage{
SdcForetakPerson(
  data,
  between = NULL,
  within = NULL,
  by = NULL,
  roundBase = 3,
  maxN = NULL,
  protectZeros = FALSE,
  secondaryZeros = FALSE,
  freqVar = NULL,
  sector = "sektor",
  private = "Privat",
  nace = c("nar*", "NACE*", "nace*"),
  nace00 = "00",
  frtk = "FRTK_ID_SSB",
  virk = "VIRK_ID_SSB",
  unik = "UNIK_ID",
  makeunik = TRUE,
  removeZeros = !protectZeros,
  preAggregate = TRUE,
  output = NULL,
  decimal = FALSE
)
}
\arguments{
\item{data}{Datasett som data frame}

\item{between}{Variabler som grupperer foretak for prikking}

\item{within}{Ytterligere variabler innen foretak som brukes til avrunding eller prikking av personer}

\item{by}{Tid eller andre variabler som deler datasettet. Metoden kjøres på hver del og resultatet settes sammen.}

\item{roundBase}{Base for avrunding}

\item{maxN}{Max-verdi for primærprikking av personer.  Ikke-NULL verdi betyr prikking istedenfor avrunding.}

\item{protectZeros}{Suppression parameter. Empty cells (count=0) are set as primary suppressed When TRUE.}

\item{secondaryZeros}{Suppression parameter.}

\item{freqVar}{A single variable holding counts (name or number) or NULL in the case of micro data}

\item{sector}{Sektor-variabel som inneholder Privat-koden (se parameter \code{"private"}).}

\item{private}{Privat-koden}

\item{nace}{between-variabel med nace-kode eller koding (starter med) som leter etter slik variabel (\code{nace} kan settes til \code{NULL})}

\item{nace00}{nace-koden som foretrekkes til sekundærprikking  (\code{nace00} kan settes til \code{NULL})}

\item{frtk}{foretak variabel}

\item{virk}{virksomhet variabel}

\item{unik}{unik variabel}

\item{makeunik}{Unik variabel genereres ved TRUE ellers antas det at den finnes}

\item{removeZeros}{When TRUE, rows with zero count will be removed from the data within the algorithm.
Default er at parameteren er motsatt av \code{protectZeros}.
Altså 0-er i data fjernes når 0-er ikke skal prikkes.
Parameteren har betydning for telling av antall foretak bak et tall.}

\item{preAggregate}{Input til \code{\link{GaussSuppressionFromData}}. Parameteren er med her for testing og sammenlikning av resultater.}

\item{output}{Ved avrunding kan ulike type output velges. Enten "rounded" (samme som NULL) eller "suppressed" (liste med begge hvis noe annet).
Her kan det bli endring.}

\item{decimal}{Ved TRUE og når maxN er ikke-NULL kjøres \code{\link{GaussSuppressDec}}. Ekstra kolloner i output er
\itemize{
\item \strong{\code{freqDec}:} Heltall og syntetiske desimaltall istedenfor prikker.
\item \strong{\code{isPublish}:} Om dette er en vanlig output-celle.
\item \strong{\code{isInner}:} Om dette er en indre celle som kan benyttes til aggregering av andre celler.
}}
}
\value{
data frame
}
\description{
Prikking av foretak og avrunding eller prikking av personer
}
\examples{

prikkeVarA <- c("arb_fylke", "ARB_ARBKOMM", "nar8", "sektor")
prikkeVarB <- c("arb_fylke", "ARB_ARBKOMM", "nar17")

z <- SdcData("syssel27")

SdcForetakPerson(z, between = prikkeVarA)
SdcForetakPerson(z, between = prikkeVarA, output = "suppressed")

SdcForetakPerson(z, between = prikkeVarB, within = "PERS_KJOENN")

z100 <- SdcData("syssel100")
out <- SdcForetakPerson(z100, between = prikkeVarB, within = c("PERS_KJOENN", "alder6"))
head(out)
tail(out)

# Lager data med to stataar
z100$stataar <- "2019"
z$stataar <- "2020"
z127 <- rbind(z100, z)
out127 <- SdcForetakPerson(z127, between = prikkeVarB, by = "stataar")
head(out127)
tail(out127)
}
