% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SdcForetakPersonDecimalTest.R
\name{SdcForetakPersonDecimalTest}
\alias{SdcForetakPersonDecimalTest}
\title{Test av SdcForetakPerson med decimal}
\usage{
SdcForetakPersonDecimalTest(data, data2, ..., stopWhenFALSE = TRUE)
}
\arguments{
\item{data}{vanlige data}

\item{data2}{data brukt når decimal er input (noen variabler kan fjernes)}

\item{...}{SdcForetakPerson-parametere}

\item{stopWhenFALSE}{Feiler ved FALSE når TRUE}
}
\value{
logical
}
\description{
Om det blir samme svar via \code{decimal = TRUE}
}
\examples{
prikkeVarA <- c("arb_fylke", "ARB_ARBKOMM", "nar8", "sektor")
z <- SdcData("syssel27")
SdcForetakPersonDecimalTest(z, z[1:7], between = prikkeVarA)
}
\keyword{internal}
