% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SdcForetakPerson.R
\name{SdcForetakPerson}
\alias{SdcForetakPerson}
\title{Prikking av foretak og avrunding eller prikking av personer}
\usage{
SdcForetakPerson(
  data,
  between = NULL,
  within = NULL,
  by = NULL,
  roundBase = 3,
  maxN = NULL,
  protectZeros = FALSE,
  secondaryZeros = FALSE,
  freqVar = NULL,
  sector = lower_match(data, "sektor"),
  private = "Privat",
  nace = c("nar*", "NACE*", "nace*"),
  nace00 = "00",
  nace00primary = FALSE,
  frtk = lower_match(data, "frtk_id_ssb"),
  virk = lower_match(data, "virk_id_ssb"),
  unik = lower_match(data, "unik_id"),
  makeunik = TRUE,
  removeZeros = !protectZeros,
  preAggregate = TRUE,
  output = NULL,
  decimal = FALSE,
  freqDec = "freqDec*",
  nRep = 3,
  digitsA = 9,
  digitsB = 5,
  allowTotal = FALSE,
  til0 = TRUE,
  iWait = Inf,
  k1 = NULL,
  k2 = 2 * k1,
  pPercent = NULL,
  use_freqVar = FALSE,
  decimal_waring = TRUE
)
}
\arguments{
\item{data}{Datasett som data frame}

\item{between}{Variabler som grupperer foretak for prikking}

\item{within}{Ytterligere variabler innen foretak som brukes til avrunding eller prikking av personer}

\item{by}{Tid eller andre variabler som deler datasettet. Metoden kjøres på hver del og resultatet settes sammen.}

\item{roundBase}{Base for avrunding}

\item{maxN}{Max-verdi for primærprikking av personer.  Ikke-NULL verdi betyr prikking istedenfor avrunding.}

\item{protectZeros}{Suppression parameter. Empty cells (count=0) are set as primary suppressed When TRUE.}

\item{secondaryZeros}{Suppression parameter.}

\item{freqVar}{A single variable holding counts (name or number) or NULL in the case of micro data (then variable named \code{"freq"} will be generated )}

\item{sector}{Sektor-variabel som inneholder Privat-koden (se parameter \code{"private"}).}

\item{private}{Privat-koden}

\item{nace}{between-variabel med nace-kode eller koding (starter med) som leter etter slik variabel (\code{nace} kan settes til \code{NULL})}

\item{nace00}{nace-koden som foretrekkes til sekundærprikking  (\code{nace00} kan settes til \code{NULL})}

\item{nace00primary}{Ved FALSE utelates nace00-koden (se over) fra primærprikking}

\item{frtk}{foretak variabel}

\item{virk}{virksomhet variabel}

\item{unik}{unik variabel}

\item{makeunik}{Unik variabel genereres ved TRUE ellers antas det at den finnes}

\item{removeZeros}{When TRUE, rows with zero count will be removed from the data within the algorithm.
Default er at parameteren er motsatt av \code{protectZeros}.
Altså 0-er i data fjernes når 0-er ikke skal prikkes.
Parameteren har betydning for telling av antall foretak bak et tall.}

\item{preAggregate}{Input til \code{\link{GaussSuppressionFromData}}. Parameteren er med her for testing og sammenlikning av resultater.}

\item{output}{Ved avrunding kan ulike type output velges. Enten "rounded" (samme som NULL) eller "suppressed" (liste med begge hvis noe annet).
Her kan det bli endring.}

\item{decimal}{\strong{Ved TRUE} returneres indre celle-data med desimaltall. Dette kan fungere som input seinere (se nedenfor).
Når, i tillegg, \code{maxN} er \code{NULL} (default), er det mulig å spesifisere \code{between} som en formel (se eksempel).

\strong{Ved \code{decimal} som en data-frame} antas at dette er indre celle-data med desimaltall. Prikking vil baseres på aggregering av disse.}

\item{freqDec}{Navn på variabel(er) med desimaltall eller koding (starter med). Brukes når \code{decimal} er en data-frame.}

\item{nRep}{Antall desimaltallsvariabler, \code{\link{GaussSuppressDec}} parameter.}

\item{digitsA}{\code{\link{GaussSuppressDec}} parameter (9 er vanligvis ok)}

\item{digitsB}{\code{\link{SuppressionFromDecimals}} parameter (5 er ok når nRep=3)}

\item{allowTotal}{Når TRUE, ingen prikking når alle within-variabler er \code{"Total"}.}

\item{til0}{Når TRUE: Når ikke-prikket tall som følge av \code{allowTotal=TRUE} er avrundet til 0 blir
prikking av undergrupper av denne opphevet og erstattet med 0.}

\item{iWait}{Minimum antall sekunder mellom hver gang det skrives ut ekstra informasjon fra prikkerutinen.}

\item{k1}{Parameter for dominansregel. Primærprikking når ett privat foretak har mer enn k1\% av antall personer i den aktuelle tabellcellen.}

\item{k2}{Parameter for dominansregel. Primærprikking når to private foretak har mer enn k2\% av antall personer i den aktuelle tabellcellen.}

\item{pPercent}{Parameter for p\%-regel som er alternativ til dominansregel. Dersom denne spesifiseres, ignoreres k1 og k2.}

\item{use_freqVar}{\code{\link[GaussSuppression]{GaussSuppressDec}} parameter.  TRUE betyr at desimaltallene er syntetiske frekvenstall.
Ved FALSE (default) varierer desimaltallene  rundt 0.}

\item{decimal_waring}{Når TRUE: Det sjekkes om aggregerte frekvenser fra desimaldata-input samsvarer med frekvenser beregnet fra data-input
(variabelen "original" i output fra \code{\link{PLSroundingSuppressed}}. Ved avvik kommer warning og det skrives ut noen rader.
Begge datasettene kan sees ved å endre parameteren \code{output} (se over).}
}
\value{
data frame
}
\description{
Prikking av foretak og avrunding eller prikking av personer.
Sett parameteren \code{allowTotal} til \code{TRUE} for at kategorier innen (\code{within}) foretak skal prikkes samtidig som totalverdier over disse grupperingene tillates publisert.
}
\details{
Default parameterverdier der funksjonen \code{\link{lower_match}} brukes betyr at default er første variabelnavn i \code{data}
som matcher når det ikke tas hensyn til små eller store bokstaver. Dette betyr, for eksempel, at parameteren \code{frtk}
ikke må bli spesifisert av brukeren hvis variabelnavnet i inputdata er \code{"frtk_id_ssb"} eller \code{"FRTK_ID_SSB"}.
Det kan også være \code{"frtk_id_SSB"} om man vil.
}
\examples{

prikkeVarA <- c("arb_fylke", "ARB_ARBKOMM", "nar8", "sektor")
prikkeVarB <- c("arb_fylke", "ARB_ARBKOMM", "nar17")

z <- SdcData("syssel27")

SdcForetakPerson(z, between = prikkeVarA)
SdcForetakPerson(z, between = prikkeVarA, output = "suppressed")

SdcForetakPerson(z, between = prikkeVarB, within = "PERS_KJOENN")

SdcForetakPerson(z, between = prikkeVarA, maxN = 2)
SdcForetakPerson(z, between = prikkeVarA, maxN = 2, decimal = TRUE)
SdcForetakPerson(z, between = prikkeVarB, within = "PERS_KJOENN", maxN = 2, decimal = TRUE)

z100 <- SdcData("syssel100")
out <- SdcForetakPerson(z100, between = prikkeVarB, within = c("PERS_KJOENN", "alder6"))
head(out)
tail(out)

# Bruke av dominansregler
outD <- SdcForetakPerson(z100, between = prikkeVarB, within = c("PERS_KJOENN", "alder6"), 
                         k1 = 40, k2 = 50, output = "suppressed")
outD                        

# Setter  allowTotal = TRUE
outT <- SdcForetakPerson(z100, between = prikkeVarB, within = c("PERS_KJOENN", "alder6"), 
                         allowTotal = TRUE)
# Rader som har gitt ulik prikking
rader <- which(is.na(outT$roundedSuppressed) != is.na(out$roundedSuppressed))
out[rader, ]
outT[rader, ]

# Ser effekt av allowTotal ved bare prikking
outP <- SdcForetakPerson(z100, between = prikkeVarB, within = c("PERS_KJOENN", "alder6"), maxN = 1)
outTP <- SdcForetakPerson(z100, between = prikkeVarB, within = c("PERS_KJOENN", "alder6"), maxN = 1, 
                          allowTotal = TRUE)
# Rader som har gitt ulik primærprikking 
raderP <- which(outTP$primary != outP$primary)
outP[raderP, ]   # Her ble allikevel prikking til slutt lik (suppressed)
outTP[raderP, ]  # Dette pga. singleton-håndtering (1-ere som kan avsløres)


# Finner data desimaltall med mange variabler som tas hensyn til.  
# Dessverre en warning som kan sees bort fra 
# Kan unngaas med dataDec <- suppressWarnings(SdcForetakPerson(.....
prikkeVarC <- c("arb_fylke", "ARB_ARBKOMM", "nar8", "sektor", "nar17")
dataDec <- SdcForetakPerson(z100, between = prikkeVarC, nace = "nar8", decimal = TRUE)

# Bruker desimaltall som utgangspunkt for prikking
outA <- SdcForetakPerson(z100, between = prikkeVarA, decimal = dataDec)
outB <- SdcForetakPerson(z100, between = prikkeVarB, within = "PERS_KJOENN", decimal = dataDec)

# Desimaltall kan genereres med formel 
dataDec2 <- SdcForetakPerson(z100, between = 
                 ~(arb_fylke + ARB_ARBKOMM) * nar8 * sektor + (arb_fylke + ARB_ARBKOMM) * nar17, 
                 nace = "nar8", decimal = TRUE)

# Lager data med to stataar
z100$stataar <- "2019"
z$stataar <- "2020"
z127 <- rbind(z100, z)
out127 <- SdcForetakPerson(z127, between = prikkeVarB, by = "stataar")
head(out127)
tail(out127)
}
