% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SdcForetakPersonDecimalTest.R
\name{SdcForetakPersonDecimalTest}
\alias{SdcForetakPersonDecimalTest}
\title{Test av SdcForetakPerson med decimal}
\usage{
SdcForetakPersonDecimalTest(...)
}
\arguments{
\item{...}{SdcForetakPerson-parametere}
}
\value{
logical
}
\description{
Om det blir samme svar via \code{decimal = TRUE}
}
\examples{
prikkeVarA <- c("arb_fylke", "ARB_ARBKOMM", "nar8", "sektor")
z <- SdcData("syssel27")
SdcForetakPersonDecimalTest(z, between = prikkeVarA)
}
\keyword{internal}
